requirements.txt
openai
autogen         
pyautogen
python-dotenv
tqdm
pandas
tabulate
transformers
torch
torchvision
torchaudio
sqlalchemy
pandas